+++ techpick/src/components/FolderAndPickDndContextProvider/DragOverlay/PickDragOverlayShadowList.tsx
