+++ techpick/src/components/FolderAndPickDndContextProvider/DragOverlay/dragOverlay.css.ts
