+++ techpick/src/components/FolderAndPickDndContextProvider/DragOverlay/folderItemOverlay.css.ts
