+++ techpick/src/components/FolderAndPickDndContextProvider/DragOverlay/pickRecordOverlay.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 import { colorVars } from 'techpick-shared';
-import { pickRecordLayoutStyle } from '../PickRecord/pickRecord.css';
+import { pickRecordLayoutStyle } from '../../PickRecord/pickRecord.css';
 
 export const pickRecordOverlayLayoutStyle = style([
   pickRecordLayoutStyle,
